pyproject.toml
[project]
name = "bifrost"
version = "0.1.0"
description = "CLI bridge between local dev and remote test setups"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9",
    "rich>=13.0",
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[project.scripts]
bf = "bifrost.cli.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bifrost"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "mypy>=1.19",
    "pre-commit>=4.0",
    "pytest>=8.0",
    "ruff>=0.15",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.ruff]
target-version = "py310"
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "F",      # pyflakes
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "SIM",    # flake8-simplify
    "RUF",    # ruff-specific
]

[tool.ruff.lint.isort]
known-first-party = ["bifrost"]

[tool.mypy]
python_version = "3.10"
mypy_path = "src"
packages = ["bifrost", "tests"]
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
